separated scripts to run the scraper and page separately. Use npm run page to run a single page and npm run scraper to run the whole scraper. (devlopment only)
