+++ scraper/package.json
@@ -4,7 +4,8 @@
   "description": "",
   "scripts": {
     "start": "ts-node writeToDb.ts",
-    "scraper": "ts-node scraper.ts",
+    "scraper": "ts-node runScraper.ts",
+    "page": "ts-node runPage.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "",
