+++ scraper/scraper.ts
@@ -97,229 +97,157 @@ const getPageUrls = async ({
  *    console.log(data) // false
  */
 const timetableScraper = async (
-    year: number
-  ): Promise<
-    { timetableData: TimetableData; courseWarnings: CourseWarning[] } | false
-  > => {
-    // Launch the browser. Headless mode = true by default
-    const browser = await puppeteer.launch({ headless: false })
-    try {
-      const batchsize = 50
-      // Create batchsize pages to scrape each course
-      const pages = await createPages({
-        browser: browser,
-        batchsize: batchsize,
-      })
-      let page = pages[0]
-      // Base url to be used for all scraping
-      const base = `http://timetable.unsw.edu.au/${year}/`
+  year: number
+): Promise<
+  { timetableData: TimetableData; courseWarnings: CourseWarning[] } | false
+> => {
+  // Launch the browser. Headless mode = true by default
+  const browser = await puppeteer.launch({ headless: false })
+  try {
+    const batchsize = 50
+    // Create batchsize pages to scrape each course
+    const pages = await createPages({
+      browser: browser,
+      batchsize: batchsize,
+    })
+    let page = pages[0]
+    // Base url to be used for all scraping
+    const base = `http://timetable.unsw.edu.au/${year}/`
 
-      // JSON Array to store the course data.
-      // There are no cases which cannot be classified. (Yet)
-      const timetableData: TimetableData = {
-        Summer: [],
-        T1: [],
-        T2: [],
-        T3: [],
-        S1: [],
-        S2: [],
-        Other: [],
+    // JSON Array to store the course data.
+    // There are no cases which cannot be classified. (Yet)
+    const timetableData: TimetableData = {
+      Summer: [],
+      T1: [],
+      T2: [],
+      T3: [],
+      S1: [],
+      S2: [],
+      Other: [],
+    }
+
+    // CourseWarning array for any fields not aligning with the strict requirements
+    const courseWarnings: CourseWarning[] = []
+
+    // Go to the page with list of subjects (Accounting, Computers etc)
+    await page.goto(base, {
+      waitUntil: 'networkidle2',
+    })
+
+    // Defining the regex for course scraping...
+    const courseUrlRegex = /([A-Z]{8})\.html/
+
+    // Gets all the dataurls on the timetable page.
+    const urlList = await getUrls({
+      page: page,
+      regex: courseUrlRegex,
+    })
+
+    // Defining the regex for each of the subject codes...
+    const subjectUrlRegex = /([A-Z]{4}[0-9]{4})\.html/
+
+    // Jobs -> urls of each subject
+    let jobs: UrlList = []
+
+    // Scrape all the urls from the subject pages (eg: COMPKENS, etc)
+    for (let url = 0; url < urlList.length; ) {
+      // List of promises that are being resolved
+      const promises: Promise<string[]>[] = []
+      // array of resolved promises from the promises array
+      let result: UrlList[] = []
+
+      // Open a different url on a different page
+      for (let i = 0; i < batchsize && url < urlList.length; i++) {
+        const urls = getPageUrls({
+          url: urlList[url],
+          page: pages[i],
+          regex: subjectUrlRegex,
+        })
+        promises.push(urls)
+        url++
       }
+      // Wait for all the pages
+      result = await Promise.all(promises)
 
-      // CourseWarning array for any fields not aligning with the strict requirements
-      const courseWarnings: CourseWarning[] = []
-
-      // Go to the page with list of subjects (Accounting, Computers etc)
-      await page.goto(base, {
-        waitUntil: 'networkidle2',
-      })
-
-      // Defining the regex for course scraping...
-      const courseUrlRegex = /([A-Z]{8})\.html/
-
-      // Gets all the dataurls on the timetable page.
-      const urlList = await getUrls({
-        page: page,
-        regex: courseUrlRegex,
-      })
-
-      // Defining the regex for each of the subject codes...
-      const subjectUrlRegex = /([A-Z]{4}[0-9]{4})\.html/
-
-      // Jobs -> urls of each subject
-      let jobs: UrlList = []
-
-      // Scrape all the urls from the subject pages (eg: COMPKENS, etc)
-      for (let url = 0; url < urlList.length; ) {
-        // List of promises that are being resolved
-        const promises: Promise<string[]>[] = []
-        // array of resolved promises from the promises array
-        let result: UrlList[] = []
-
-        // Open a different url on a different page
-        for (let i = 0; i < batchsize && url < urlList.length; i++) {
-          const urls = getPageUrls({
-            url: urlList[url],
-            page: pages[i],
-            regex: subjectUrlRegex,
-          })
-          promises.push(urls)
-          url++
-        }
-        // Wait for all the pages
-        result = await Promise.all(promises)
-
-        // Then add them to the jobs queue
-        for (const pageurls of result) {
-          jobs = jobs.concat(pageurls)
-        }
+      // Then add them to the jobs queue
+      for (const pageurls of result) {
+        jobs = jobs.concat(pageurls)
       }
+    }
 
-      // Now scrape each page in the jobs queue and then add it to the scraped
-      // array
-      for (let job = 0; job < jobs.length; ) {
-        const promises: Promise<{
-          coursesData: TimetableData
-          courseWarnings: CourseWarning[]
-        }>[] = []
-        let result: {
-          coursesData: TimetableData
-          courseWarnings: CourseWarning[]
-        }[]
-        for (let i = 0; i < batchsize && job < jobs.length; i++) {
-          const data = scrapeSubject({
-            page: pages[i],
-            course: jobs[job],
-          })
-          promises.push(data)
-          job++
-        }
-        // Wait for all the pages
-        result = await Promise.all(promises)
+    // Now scrape each page in the jobs queue and then add it to the scraped
+    // array
+    for (let job = 0; job < jobs.length; ) {
+      const promises: Promise<{
+        coursesData: TimetableData
+        courseWarnings: CourseWarning[]
+      }>[] = []
+      let result: {
+        coursesData: TimetableData
+        courseWarnings: CourseWarning[]
+      }[]
+      for (let i = 0; i < batchsize && job < jobs.length; i++) {
+        const data = scrapeSubject({
+          page: pages[i],
+          course: jobs[job],
+        })
+        promises.push(data)
+        job++
+      }
+      // Wait for all the pages
+      result = await Promise.all(promises)
 
-        for (const element of result) {
-          courseWarnings.push(...element.courseWarnings)
-          for (const scrapedTerm of keysOf(element.coursesData)) {
-            // Each termlist may contain multiple courses
-            for (const scrapedCourse of element.coursesData[scrapedTerm]) {
-              if (!scrapedCourse) {
-                console.log(element, 'Error in result: ', result)
-                throw new Error()
+      for (const element of result) {
+        courseWarnings.push(...element.courseWarnings)
+        for (const scrapedTerm of keysOf(element.coursesData)) {
+          // Each termlist may contain multiple courses
+          for (const scrapedCourse of element.coursesData[scrapedTerm]) {
+            if (!scrapedCourse) {
+              console.log(element, 'Error in result: ', result)
+              throw new Error()
+            }
+            // If the term is in the other list, then it has no classes. Classify it!
+            if (scrapedTerm === OtherTerms.Other) {
+              const termlist: ExtendedTerm[] = getTermFromCourse({
+                course: scrapedCourse,
+              })
+              if (termlist == []) {
+                termlist.push(OtherTerms.Other)
               }
-              // If the term is in the other list, then it has no classes. Classify it!
-              if (scrapedTerm === OtherTerms.Other) {
-                const termlist: ExtendedTerm[] = getTermFromCourse({
-                  course: scrapedCourse,
+              const notes = scrapedCourse.notes
+              let noteIndex = 0
+              for (const term of termlist) {
+                if (notes?.[noteIndex] && notes[noteIndex] != ' ') {
+                  scrapedCourse.notes = [notes[noteIndex]]
+                } else {
+                  delete scrapedCourse.notes
+                }
+
+                timetableData[term].push({
+                  ...cloneDeep(scrapedCourse),
                 })
-                if (termlist == []) {
-                  termlist.push(OtherTerms.Other)
-                }
-                const notes = scrapedCourse.notes
-                let noteIndex = 0
-                for (const term of termlist) {
-                  if (notes?.[noteIndex] && notes[noteIndex] != ' ') {
-                    scrapedCourse.notes = [notes[noteIndex]]
-                  } else {
-                    delete scrapedCourse.notes
-                  }
-
-                  timetableData[term].push({
-                    ...cloneDeep(scrapedCourse),
-                  })
-                }
-              } else {
-                // If the term is already assigned then just add it to the respective term.
-                timetableData[scrapedTerm].push(scrapedCourse)
               }
+            } else {
+              // If the term is already assigned then just add it to the respective term.
+              timetableData[scrapedTerm].push(scrapedCourse)
             }
           }
         }
       }
-
-      // Close the browser.
-      return { timetableData: timetableData, courseWarnings: courseWarnings }
-    } catch (err) {
-      // log error and close browser.
-      console.error(err)
-      return false
-    } finally {
-      await browser.close()
-    }
-  }
-
-  // This function is for devlopment purposes only. Use npm run scraper to test the scraper
-;(async () => {
-  console.time('cscraper')
-  // const browser = await puppeteer.launch({ headless: false })
-  try {
-    const data = await timetableScraper(2020)
-
-    if (data === false) {
-      return
     }
 
-    const fs = require('fs')
-    fs.writeFile(
-      'Other.json',
-      JSON.stringify(data.timetableData.Other),
-      'utf-8',
-      (err: unknown) => {
-        if (err) {
-          console.error(err)
-        }
-      }
-    )
-
-    fs.writeFile(
-      'T1.json',
-      JSON.stringify(data.timetableData.T1),
-      'utf-8',
-      (err: unknown) => {
-        if (err) {
-          console.error(err)
-        }
-      }
-    )
-
-    // const singlepage = await browser.newPage()
-    // const data = await scrapeSubject({
-    //   page: singlepage,
-    //   course: 'http://timetable.unsw.edu.au/2019/ACCT5914.html',
-    // })
-
-    // const fs = require('fs')
-    // fs.writeFile(
-    //   'T1.json',
-    //   JSON.stringify(data.coursesData.T1),
-    //   'utf-8',
-    //   (err: unknown) => {
-    //     if (err) {
-    //       console.error(err)
-    //     }
-    //   }
-    // )
-    // fs.writeFile(
-    //   'T1Warn.json',
-    //   JSON.stringify(data.courseWarnings),
-    //   'utf-8',
-    //   (err: unknown) => {
-    //     if (err) {
-    //       console.error(err)
-    //     }
-    //   }
-    // )
+    // Close the browser.
+    return { timetableData: timetableData, courseWarnings: courseWarnings }
   } catch (err) {
-    console.log(err)
+    // log error and close browser.
+    console.error(err)
+    return false
   } finally {
-    // await browser.close()
+    await browser.close()
   }
-  console.timeEnd('cscraper')
-  const used = process.memoryUsage()
-  for (let key in used) {
-    console.log(
-      `${key} ${Math.round((used[key] / 1024 / 1024) * 100) / 100} MB`
-    )
-  }
-})()
+}
 
 export { timetableScraper }
+
+// Development only
+export { scrapeSubject }
